requirements.txt
streamlit
pandas
altair
numpy
python-pptx
pillow